extract common cudnn fusion utils
